Improvements to sequence handling in Expression implementations and updates to the use of ExpressionMatrix in CollectionExpression implementations.

+++ code/src/main/java/io/almostrealism/expression/Equals.java
@@ -51,7 +51,7 @@ public class Equals extends Comparison {
 	}
 
 	@Override
-	protected IndexSequence compare(IndexSequence left, IndexSequence right, int len) {
+	protected IndexSequence compare(IndexSequence left, IndexSequence right, long len) {
 		return left.eq(right);
 	}
